Macro codegen::proto

proto!() { /* proc-macro */ }

Generates a protocol definition.

This macro generates hardcoded symbols like Client, Proto and Server. Each codegen::proto! should have its dedicated module.

Syntax

The macro block can start with inner attributes, which are going to be applied on the generated Proto struct.

The content then starts with the protocol name: name = "xxx";.

Then there can be any number of messages of queries, with the following format:

ITEM := OUTER_ATTRIBUTES* (MESSAGE | QUERY)

# The IDENT is the message name.
# The FIELD_LIST contains the fields of the message.
MESSAGE := DIRECTION "message" IDENT FIELD_LIST

# The IDENT is the query name. The request and response structs start with the query name,
followed by "Request" or "Response".
# The first FIELD_LIST contains the fields of the request.
# The second FIELD_LIST contains the fields of the request.
QUERY := DIRECTION "query" IDENT FIELD_LIST "->" FIELD_LIST

# DIRECTION indicates the *source* of the message or the *sender* of the query request
(i.e. the *receiver* of the query *response*)
DIRECTION := "client" | "server" | "mutual"

FIELD_LIST := "{" ( FIELD )* "}"

# The outer attributes are applied to the message struct, or the request struct for queries
# The IDENT is the name of the field.
# The TYPE is the type of the field.
FIELD := OUTER_ATTRIBUTES* IDENT ":" TYPE ","