Struct codegen::Enum

pub struct Enum { /* fields omitted */ }

Defines an enumeration.

Methods

impl Enum

pub fn new(name: &str) -> Self

Return a enum definition with the provided name.

pub fn ty(&self) -> &Type

Returns a reference to the type.

pub fn vis(&mut self, vis: &str) -> &mut Self

Set the enum visibility.

pub fn generic(&mut self, name: &str) -> &mut Self

Add a generic to the enum.

pub fn bound<T>(&mut self, name: &str, ty: T) -> &mut Self where
T: Into<Type>,

Add a where bound to the enum.

pub fn doc(&mut self, docs: &str) -> &mut Self

Set the enum documentation.

pub fn derive(&mut self, name: &str) -> &mut Self

Add a new type that the struct should derive.

pub fn new_variant(&mut self, name: &str) -> &mut Variant

Push a variant to the enum, returning a mutable reference to it.

pub fn push_variant(&mut self, item: Variant) -> &mut Self

Push a variant to the enum.

pub fn fmt(&self, fmt: &mut Formatter) -> Result

Formats the enum using the given formatter.