Struct codegen::Impl

pub struct Impl { /* fields omitted */ }

Defines an impl block.

Methods

impl Impl

pub fn new<T>(target: T) -> Self where
T: Into<Type>,

Return a new impl definition

pub fn generic(&mut self, name: &str) -> &mut Self

Add a generic to the impl block.

This adds the generic for the block (impl<T>) and not the target type.

pub fn target_generic<T>(&mut self, ty: T) -> &mut Self where
T: Into<Type>,

Add a generic to the target type.

pub fn impl_trait<T>(&mut self, ty: T) -> &mut Self where
T: Into<Type>,

Set the trait that the impl block is implementing.

pub fn associate_type<T>(&mut self, name: &str, ty: T) -> &mut Self where
T: Into<Type>,

Set an associated type.

pub fn bound<T>(&mut self, name: &str, ty: T) -> &mut Self where
T: Into<Type>,

Add a where bound to the impl block.

pub fn new_fn(&mut self, name: &str) -> &mut Function

Push a new function definition, returning a mutable reference to it.

pub fn push_fn(&mut self, item: Function) -> &mut Self

Push a function definition.

pub fn fmt(&self, fmt: &mut Formatter) -> Result

Formats the impl block using the given formatter.