Struct codegen::Scope

pub struct Scope { /* fields omitted */ }

Defines a scope.

A scope contains modules, types, etc...

Methods

impl Scope

pub fn new() -> Self

Returns a new scope

pub fn import(&mut self, path: &str, ty: &str) -> &mut Import

Import a type into the scope.

This results in a new use statement being added to the beginning of the scope.

pub fn new_module(&mut self, name: &str) -> &mut Module

Push a new module definition, returning a mutable reference to it.

Panics

Since a module's name must uniquely identify it within the scope in which it is defined, pushing a module whose name is already defined in this scope will cause this function to panic.

In many cases, the get_or_new_module function is preferrable, as it will return the existing definition instead.

pub fn get_module_mut<Q: ?Sized>(&mut self, name: &Q) -> Option<&mut Module> where
String: PartialEq<Q>,

Returns a mutable reference to a module if it is exists in this scope.

pub fn get_module<Q: ?Sized>(&self, name: &Q) -> Option<&Module> where
String: PartialEq<Q>,

Returns a mutable reference to a module if it is exists in this scope.

pub fn get_or_new_module(&mut self, name: &str) -> &mut Module

Returns a mutable reference to a module, creating it if it does not exist.

pub fn push_module(&mut self, item: Module) -> &mut Self

Push a module definition.

Panics

Since a module's name must uniquely identify it within the scope in which it is defined, pushing a module whose name is already defined in this scope will cause this function to panic.

In many cases, the get_or_new_module function is preferrable, as it will return the existing definition instead.

pub fn new_struct(&mut self, name: &str) -> &mut Struct

Push a new struct definition, returning a mutable reference to it.

pub fn push_struct(&mut self, item: Struct) -> &mut Self

Push a struct definition

pub fn new_trait(&mut self, name: &str) -> &mut Trait

Push a new trait definition, returning a mutable reference to it.

pub fn push_trait(&mut self, item: Trait) -> &mut Self

Push a trait definition

pub fn new_enum(&mut self, name: &str) -> &mut Enum

Push a new struct definition, returning a mutable reference to it.

pub fn push_enum(&mut self, item: Enum) -> &mut Self

Push a structure definition

pub fn new_impl(&mut self, target: &str) -> &mut Impl

Push a new impl block, returning a mutable reference to it.

pub fn push_impl(&mut self, item: Impl) -> &mut Self

Push an impl block.

pub fn raw(&mut self, val: &str) -> &mut Self

Push a raw string to the scope.

This string will be included verbatim in the formatted string.

pub fn to_string(&self) -> String

Return a string representation of the scope.

pub fn fmt(&self, fmt: &mut Formatter) -> Result

Formats the scope using the given formatter.