Struct codegen::Struct

pub struct Struct { /* fields omitted */ }

Defines a struct.

Methods

impl Struct

pub fn new(name: &str) -> Self

Return a structure definition with the provided name

pub fn ty(&self) -> &Type

Returns a reference to the type

pub fn vis(&mut self, vis: &str) -> &mut Self

Set the structure visibility.

pub fn generic(&mut self, name: &str) -> &mut Self

Add a generic to the struct.

pub fn bound<T>(&mut self, name: &str, ty: T) -> &mut Self where
T: Into<Type>,

Add a where bound to the struct.

pub fn doc(&mut self, docs: &str) -> &mut Self

Set the structure documentation.

pub fn derive(&mut self, name: &str) -> &mut Self

Add a new type that the struct should derive.

pub fn field<T>(&mut self, name: &str, ty: T) -> &mut Self where
T: Into<Type>,

Add a named field to the struct.

A struct can either set named fields with this function or tuple fields with tuple_field, but not both.

pub fn tuple_field<T>(&mut self, ty: T) -> &mut Self where
T: Into<Type>,

Add a tuple field to the struct.

A struct can either set tuple fields with this function or named fields with field, but not both.

pub fn fmt(&self, fmt: &mut Formatter) -> Result

Formats the struct using the given formatter.